requirements.txt
oracledb
python-dotenv
tabulate
